Cleanup files; use types from core
