+++ controller/apps/rides-examiner/src/analyzers/at-most-two-driver-ids.analyzer.ts
@@ -1,14 +1,7 @@
 /* * */
 
 import { type AnalysisData } from '@/types/analysis-data.type.js';
-import { type RideAnalysis } from '@tmlmobilidade/types';
-
-/* * */
-
-interface ExplicitRideAnalysis extends RideAnalysis {
-	reason: 'FOUND_MORE_THAN_2_DRIVER_IDS' | 'FOUND_ONE_OR_TWO_DRIVER_IDS' | 'NO_DRIVER_ID_FOUND'
-	unit: 'UNIQUE_DRIVER_IDS'
-};
+import { type Ride } from '@tmlmobilidade/types';
 
 /**
  * This analyzer tests if the trip has at most two drivers (at least one, maximum of two).
@@ -17,7 +10,7 @@ interface ExplicitRideAnalysis extends RideAnalysis {
  * → PASS = At least one Driver, and maximum two Driver IDs for the trip.
  * → FAIL = No Driver or more than two Drivers IDs for the trip.
  */
-export function atMostTwoDriverIdsAnalyzer(analysisData: AnalysisData): ExplicitRideAnalysis {
+export function atMostTwoDriverIdsAnalyzer(analysisData: AnalysisData): Ride['analysis']['AT_MOST_TWO_DRIVER_IDS'] {
 	try {
 		//
 
@@ -26,7 +19,6 @@ export function atMostTwoDriverIdsAnalyzer(analysisData: AnalysisData): Explicit
 				grade: 'fail',
 				message: 'No Driver IDs found for this trip.',
 				reason: 'NO_DRIVER_ID_FOUND',
-				unit: 'UNIQUE_DRIVER_IDS',
 				value: 0,
 			};
 		}
@@ -36,7 +28,6 @@ export function atMostTwoDriverIdsAnalyzer(analysisData: AnalysisData): Explicit
 				grade: 'fail',
 				message: `Found ${analysisData.ride.driver_ids.length} Driver IDs for this trip.`,
 				reason: 'FOUND_MORE_THAN_2_DRIVER_IDS',
-				unit: 'UNIQUE_DRIVER_IDS',
 				value: analysisData.ride.driver_ids.length,
 			};
 		}
@@ -45,7 +36,6 @@ export function atMostTwoDriverIdsAnalyzer(analysisData: AnalysisData): Explicit
 			grade: 'pass',
 			message: `Found ${analysisData.ride.driver_ids.length} Driver IDs for this trip.`,
 			reason: 'FOUND_ONE_OR_TWO_DRIVER_IDS',
-			unit: 'UNIQUE_DRIVER_IDS',
 			value: analysisData.ride.driver_ids.length,
 		};
 
@@ -56,7 +46,6 @@ export function atMostTwoDriverIdsAnalyzer(analysisData: AnalysisData): Explicit
 			grade: 'error',
 			message: error.message,
 			reason: null,
-			unit: null,
 			value: null,
 		};
 	}
