+++ controller/apps/rides-examiner/src/analyzers/at-most-two-vehicle-ids.analyzer.ts
@@ -1,14 +1,7 @@
 /* * */
 
 import { type AnalysisData } from '@/types/analysis-data.type.js';
-import { type RideAnalysis } from '@tmlmobilidade/types';
-
-/* * */
-
-interface ExplicitRideAnalysis extends RideAnalysis {
-	reason: 'FOUND_MORE_THAN_2_VEHICLE_IDS' | 'FOUND_ONE_OR_TWO_VEHICLE_IDS' | 'NO_VEHICLE_ID_FOUND'
-	unit: 'UNIQUE_VEHICLE_IDS'
-};
+import { type Ride } from '@tmlmobilidade/types';
 
 /**
  * This analyzer tests if the trip has at most two vehicle IDs (at least one, maximum of two).
@@ -17,7 +10,7 @@ interface ExplicitRideAnalysis extends RideAnalysis {
  * → PASS = At least one Vehicle, and maximum two Vehicle IDs for the trip.
  * → FAIL = No Vehicle or more than two Vehicle IDs for the trip.
  */
-export function atMostTwoVehicleIdsAnalyzer(analysisData: AnalysisData): ExplicitRideAnalysis {
+export function atMostTwoVehicleIdsAnalyzer(analysisData: AnalysisData): Ride['analysis']['AT_MOST_TWO_VEHICLE_IDS'] {
 	try {
 		//
 
@@ -26,7 +19,6 @@ export function atMostTwoVehicleIdsAnalyzer(analysisData: AnalysisData): Explici
 				grade: 'fail',
 				message: 'No Vehicle IDs found for this trip.',
 				reason: 'NO_VEHICLE_ID_FOUND',
-				unit: 'UNIQUE_VEHICLE_IDS',
 				value: 0,
 			};
 		}
@@ -36,7 +28,6 @@ export function atMostTwoVehicleIdsAnalyzer(analysisData: AnalysisData): Explici
 				grade: 'fail',
 				message: `Found ${analysisData.ride.vehicle_ids.length} Vehicle IDs for this trip.`,
 				reason: 'FOUND_MORE_THAN_2_VEHICLE_IDS',
-				unit: 'UNIQUE_VEHICLE_IDS',
 				value: analysisData.ride.vehicle_ids.length,
 			};
 		}
@@ -45,7 +36,6 @@ export function atMostTwoVehicleIdsAnalyzer(analysisData: AnalysisData): Explici
 			grade: 'pass',
 			message: `Found ${analysisData.ride.vehicle_ids.length} Vehicle IDs for this trip.`,
 			reason: 'FOUND_ONE_OR_TWO_VEHICLE_IDS',
-			unit: 'UNIQUE_VEHICLE_IDS',
 			value: analysisData.ride.vehicle_ids.length,
 		};
 
@@ -56,7 +46,6 @@ export function atMostTwoVehicleIdsAnalyzer(analysisData: AnalysisData): Explici
 			grade: 'error',
 			message: error.message,
 			reason: null,
-			unit: null,
 			value: null,
 		};
 	}
