+++ controller/apps/rides-examiner/src/analyzers/excessive-vehicle-event-delay.analyzer.ts
@@ -1,14 +1,7 @@
 /* * */
 
 import { type AnalysisData } from '@/types/analysis-data.type.js';
-import { type RideAnalysis } from '@tmlmobilidade/types';
-
-/* * */
-
-interface ExplicitRideAnalysis extends RideAnalysis {
-	reason: 'ALL_VEHICLE_EVENTS_ARE_WITHIN_DELAY_LIMITS' | `THERE_ARE_${number}_VEHICLE_EVENTS_WITH_EXCESSIVE_DELAY`
-	unit: 'COUNT_OF_VEHICLE_EVENTS_WITH_EXCESSIVE_DELAY'
-};
+import { type Ride } from '@tmlmobilidade/types';
 
 /**
  * This analyzer tests if there are events with excessive delay between the vehicle and PCGI.
@@ -17,7 +10,7 @@ interface ExplicitRideAnalysis extends RideAnalysis {
  * → PASS = Delay between Vehicle and PCGI timestamps is less than 10 seconds.
  * → FAIL = Delay between Vehicle and PCGI timestamps is equal to or higher than 10 seconds.
  */
-export function excessiveVehicleEventDelayAnalyzer(analysisData: AnalysisData): ExplicitRideAnalysis {
+export function excessiveVehicleEventDelayAnalyzer(analysisData: AnalysisData): Ride['analysis']['EXCESSIVE_VEHICLE_EVENT_DELAY'] {
 	try {
 		//
 
@@ -44,7 +37,6 @@ export function excessiveVehicleEventDelayAnalyzer(analysisData: AnalysisData): 
 				grade: 'pass',
 				message: 'All vehicle events are within delay limits.',
 				reason: 'ALL_VEHICLE_EVENTS_ARE_WITHIN_DELAY_LIMITS',
-				unit: 'COUNT_OF_VEHICLE_EVENTS_WITH_EXCESSIVE_DELAY',
 				value: 0,
 			};
 		}
@@ -52,8 +44,7 @@ export function excessiveVehicleEventDelayAnalyzer(analysisData: AnalysisData): 
 		return {
 			grade: 'fail',
 			message: `Found ${countOfEventsWithDelay} vehicle events with excessive delay.`,
-			reason: `THERE_ARE_${countOfEventsWithDelay}_VEHICLE_EVENTS_WITH_EXCESSIVE_DELAY`,
-			unit: 'COUNT_OF_VEHICLE_EVENTS_WITH_EXCESSIVE_DELAY',
+			reason: `THERE_ARE_VEHICLE_EVENTS_WITH_EXCESSIVE_DELAY`,
 			value: countOfEventsWithDelay,
 		};
 
@@ -64,7 +55,6 @@ export function excessiveVehicleEventDelayAnalyzer(analysisData: AnalysisData): 
 			grade: 'error',
 			message: error.message,
 			reason: null,
-			unit: null,
 			value: null,
 		};
 	}
