+++ controller/apps/rides-examiner/src/analyzers/simple-one-vehicle-event-or-validation-transaction.analyzer.ts
@@ -1,13 +1,7 @@
 /* * */
 
 import { type AnalysisData } from '@/types/analysis-data.type.js';
-import { type RideAnalysis } from '@tmlmobilidade/types';
-
-/* * */
-
-interface ExplicitRideAnalysis extends RideAnalysis {
-	reason: 'FOUND_VEHICLE_EVENT_OR_VALIDATION_TRANSACTION' | 'NO_VEHICLE_EVENT_OR_VALIDATION_TRANSACTION_FOUND'
-};
+import { type Ride } from '@tmlmobilidade/types';
 
 /**
  * This analyzer tests if at least one vehicle event or one validation is found for the trip.
@@ -16,7 +10,7 @@ interface ExplicitRideAnalysis extends RideAnalysis {
  * → PASS = At least one Vehicle Event OR one Validation Transaction is found for the trip.
  * → FAIL = No Vehicle Events OR Validation Transactions found for the trip.
  */
-export function simpleOneVehicleEventOrValidationTransactionAnalyzer(analysisData: AnalysisData): ExplicitRideAnalysis {
+export function simpleOneVehicleEventOrValidationTransactionAnalyzer(analysisData: AnalysisData): Ride['analysis']['SIMPLE_ONE_VEHICLE_EVENT_OR_VALIDATION_TRANSACTION'] {
 	try {
 		//
 
@@ -28,7 +22,6 @@ export function simpleOneVehicleEventOrValidationTransactionAnalyzer(analysisDat
 				grade: 'pass',
 				message: `Found ${analysisData.vehicle_events.length} Vehicle Events and ${analysisData.simplified_apex_validations.length} Validation Transactions for this trip.`,
 				reason: 'FOUND_VEHICLE_EVENT_OR_VALIDATION_TRANSACTION',
-				unit: null,
 				value: null,
 			};
 		}
@@ -37,7 +30,6 @@ export function simpleOneVehicleEventOrValidationTransactionAnalyzer(analysisDat
 			grade: 'fail',
 			message: 'No Vehicle Events or Validation Transactions found for this trip.',
 			reason: 'NO_VEHICLE_EVENT_OR_VALIDATION_TRANSACTION_FOUND',
-			unit: null,
 			value: null,
 		};
 
@@ -48,7 +40,6 @@ export function simpleOneVehicleEventOrValidationTransactionAnalyzer(analysisDat
 			grade: 'error',
 			message: error.message,
 			reason: null,
-			unit: null,
 			value: null,
 		};
 	}
