+++ rides/apps/frontend/src/app/[ride_id]/layout.tsx
@@ -1,6 +1,6 @@
 /* * */
 
-import { RidesDetailContextProvider } from '@/contexts/RidesDetail.context';
+import { type RidesDetailContextProvider } from '@/contexts/RidesDetail.context';
 
 /* * */
