+++ rides/apps/frontend/src/app/[ride_id]/page.tsx
@@ -2,7 +2,7 @@
 
 /* * */
 
-import { RidesDetail } from '@/components/detail/RidesDetail';
+import { type RidesDetail } from '@/components/detail/RidesDetail';
 
 /* * */
