+++ rides/apps/frontend/src/app/board/page.tsx
@@ -2,7 +2,7 @@
 
 /* * */
 
-import { RidesBoard } from '@/components/RidesBoard';
+import { type RidesBoard } from '@/components/RidesBoard';
 
 /* * */
