+++ rides/apps/frontend/src/app/layout.tsx
@@ -1,7 +1,7 @@
 /* * */
 
 import { Providers } from '@/app/providers';
-import { RidesCatalogContextProvider } from '@/contexts/RidesCatalog.context';
+import { type RidesCatalogContextProvider } from '@/contexts/RidesCatalog.context';
 import { availableFormats } from '@/i18n/config';
 import { NavigationProgress } from '@mantine/nprogress';
 import { AppProvider, AppWrapper } from '@tmlmobilidade/ui';
