+++ rides/apps/frontend/src/app/providers.tsx
@@ -4,7 +4,7 @@
 
 import { MapOptionsContextProvider } from '@/contexts/MapOptions.context';
 import { OperationalDateContextProvider } from '@/contexts/OperationalDate.context';
-import { RidesContextProvider } from '@/contexts/Rides.context';
+import { type RidesContextProvider } from '@/contexts/Rides.context';
 import { MapProvider } from '@vis.gl/react-maplibre';
 import { PropsWithChildren } from 'react';
 import { SWRConfig, type SWRConfiguration } from 'swr';
