+++ rides/apps/frontend/src/components/detail/RidesDetail/index.tsx
@@ -2,16 +2,16 @@
 
 /* * */
 
-import { RidesDetailAnalysisResult } from '@/components/detail/RidesDetailAnalysisResult';
-import { RidesDetailAudits } from '@/components/detail/RidesDetailAudits';
-import { RidesDetailDebug } from '@/components/detail/RidesDetailDebug';
-import { RidesDetailHeader } from '@/components/detail/RidesDetailHeader';
-import { RidesDetailJustifications } from '@/components/detail/RidesDetailJustifications';
-import { RidesDetailMap } from '@/components/detail/RidesDetailMap';
-import { RidesDetailMetadata } from '@/components/detail/RidesDetailMetadata';
-import { RidesDetailPerformance } from '@/components/detail/RidesDetailPerformance';
-import { RidesDetailStops } from '@/components/detail/RidesDetailStops';
-import { RidesDetailSupport } from '@/components/detail/RidesDetailSupport';
+import { type RidesDetailAnalysisResult } from '@/components/detail/RidesDetailAnalysisResult';
+import { type RidesDetailAudits } from '@/components/detail/RidesDetailAudits';
+import { type RidesDetailDebug } from '@/components/detail/RidesDetailDebug';
+import { type RidesDetailHeader } from '@/components/detail/RidesDetailHeader';
+import { type RidesDetailJustifications } from '@/components/detail/RidesDetailJustifications';
+import { type RidesDetailMap } from '@/components/detail/RidesDetailMap';
+import { type RidesDetailMetadata } from '@/components/detail/RidesDetailMetadata';
+import { type RidesDetailPerformance } from '@/components/detail/RidesDetailPerformance';
+import { type RidesDetailStops } from '@/components/detail/RidesDetailStops';
+import { type RidesDetailSupport } from '@/components/detail/RidesDetailSupport';
 import { Pane } from '@tmlmobilidade/ui';
 
 /* * */
